move ShapeFormat to ShapeReader and ShapeWriter
