+++ src/main/java/com/spatial4j/core/io/GeoJSONReader.java
@@ -19,6 +19,7 @@ package com.spatial4j.core.io;
 
 import java.io.IOException;
 import java.io.Reader;
+import java.io.StringReader;
 import java.io.Writer;
 import java.text.NumberFormat;
 import java.text.ParseException;
@@ -44,16 +45,16 @@ import com.spatial4j.core.shape.impl.GeoCircle;
 
 
 
-public class GeoJSONFormat extends BaseFormat {
-  public static final String FORMAT = "GeoJSON";
+public class GeoJSONReader implements ShapeReader {
+  final SpatialContext ctx;
   
-  public GeoJSONFormat(SpatialContext ctx, SpatialContextFactory factory) {
-    super(ctx);
+  public GeoJSONReader(SpatialContext ctx, SpatialContextFactory factory) {
+    this.ctx = ctx;
   }
   
   @Override
   public String getFormatName() {
-    return FORMAT;
+    return ShapeIO.GeoJSON;
   }
   
   @Override
@@ -62,8 +63,15 @@ public class GeoJSONFormat extends BaseFormat {
   }
 
   @Override
-  public boolean formatMatchs(String v) {
-    return v.startsWith("{")&&v.endsWith("}");
+  public Shape read(Object value, boolean error) throws IOException, ParseException {
+    String v = value.toString().trim();
+    if(!(v.startsWith("{")&&v.endsWith("}"))) {
+      if(error) {
+        throw new ParseException("Invalid JSON", 0);
+      }
+      return null;
+    }
+    return read(new StringReader(v));
   }
   
   //--------------------------------------------------------------
@@ -327,114 +335,5 @@ public class GeoJSONFormat extends BaseFormat {
   
   protected Shape makeShapeFromCoords(String type, List coords) {
     return null;  // default is unsupported
-  }
-
-  protected void write(Writer output, NumberFormat nf, double ... coords) throws IOException {
-    output.write('[');
-    for(int i=0;i<coords.length; i++) {
-      if (i>0) {
-        output.append(',');
-      }
-      output.append(nf.format(coords[i]));
-    }
-    output.write(']');
-  }
-  
-  @Override
-  public void write(Writer output, Shape shape) throws IOException 
-  {
-    if (shape==null) {
-      throw new NullPointerException("Shape can not be null");
-    }
-    NumberFormat nf = LegacyShapeReadWriterFormat.makeNumberFormat(6);
-    if (shape instanceof Point) {
-      Point v = (Point)shape;
-      output.append("{\"type\":\"Point\",\"coordinates\":");
-      write(output, nf, v.getX(), v.getY());
-      output.append('}');
-      return;
-    }
-    if (shape instanceof Rectangle) {
-      Rectangle v = (Rectangle)shape;
-      output.append("{\"type\":\"Polygon\",\"coordinates\": [[");
-      write(output, nf, v.getMinX(), v.getMinY()); output.append(',');
-      write(output, nf, v.getMinX(), v.getMaxY()); output.append(',');
-      write(output, nf, v.getMaxX(), v.getMaxY()); output.append(',');
-      write(output, nf, v.getMaxX(), v.getMinY()); 
-      output.append("]]}");
-      return;
-    }
-    if (shape instanceof BufferedLine) {
-      BufferedLine v = (BufferedLine)shape;
-      output.append("{\"type\":\"LineString\",\"coordinates\": [");
-      write(output, nf, v.getA().getX(), v.getA().getY()); output.append(',');
-      write(output, nf, v.getB().getX(), v.getB().getY()); output.append(',');
-      output.append("]");
-      if (v.getBuf()>0) {
-        output.append("\"buffer\":");
-        output.append(nf.format(v.getBuf()));
-      }
-      output.append('}');
-      return;
-    }
-    if (shape instanceof BufferedLineString) {
-      BufferedLineString v = (BufferedLineString)shape;
-      output.append("{\"type\":\"LineString\",\"coordinates\": [");
-      BufferedLine last = null;
-      Iterator<BufferedLine> iter = v.getSegments().iterator();
-      while(iter.hasNext()) {
-        BufferedLine seg = iter.next();
-        if (last!=null) {
-          output.append(',');
-        }
-        write(output, nf, seg.getA().getX(), seg.getA().getY());
-        last = seg;
-      }
-      if (last!=null) {
-        output.append(',');
-        write(output, nf, last.getB().getX(), last.getB().getY());
-      }
-      output.append("]");
-      if (v.getBuf()>0) {
-        output.append("\"buffer\":");
-        output.append(nf.format(v.getBuf()));
-      }
-      output.append('}');
-      return;
-    }
-    if (shape instanceof Circle) {
-      // See: https://github.com/geojson/geojson-spec/wiki/Proposal---Circles-and-Ellipses-Geoms
-      Circle v = (Circle)shape;
-      Point center = v.getCenter();
-      output.append("{\"type\":\"Circle\",\"coordinates\":");
-      write(output, nf, center.getX(), center.getY());
-      output.append("\"radius\":");
-      if (v instanceof GeoCircle) {
-        double distKm = DistanceUtils.degrees2Dist(v.getRadius(),  DistanceUtils.EARTH_MEAN_RADIUS_KM);
-        output.append(nf.format(distKm));
-        output.append(",\"properties\":{");
-        output.append(",\"radius_units\":\"km\"}}");
-      }
-      else {
-        output.append(nf.format(v.getRadius())).append('}');
-      }
-      return;
-    }
-    if (shape instanceof ShapeCollection) {
-      ShapeCollection v = (ShapeCollection)shape;
-      output.append("{\"type\":\"GeometryCollection\",\"geometries\": [");
-      for(int i=0; i<v.size(); i++) {
-        if (i>0) {
-          output.append(',');
-        }
-        write(output, v.get(i));
-      }
-      output.append("]}");
-      return;
-    }
-    
-    output.append("{\"type\":\"Unknown\",\"wkt\":\"");
-    output.append( LegacyShapeReadWriterFormat.writeShape(shape));
-    output.append("\"}");
   }
 }
