+++ src/main/java/com/spatial4j/core/io/ShapeWriter.java
@@ -17,25 +17,23 @@
 
 package com.spatial4j.core.io;
 
+import java.io.IOException;
+import java.io.Reader;
+import java.io.Writer;
+import java.text.ParseException;
 
-import com.spatial4j.core.context.SpatialContext;
-import com.spatial4j.core.context.SpatialContextFactory;
-import com.spatial4j.core.shape.Point;
 import com.spatial4j.core.shape.Shape;
 
-import java.text.ParseException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 
-/**
- * 
- */
-@Deprecated
-public class WktShapeParser extends WKTReader {
+public interface ShapeWriter extends ShapeIO {
+  
+  /**
+   * Write a shape to the output writer
+   */
+  public void write(Writer output, Shape shape) throws IOException;
 
-  /** This constructor is required by {@link com.spatial4j.core.context.SpatialContextFactory#makeWktShapeParser(com.spatial4j.core.context.SpatialContext)}. */
-  public WktShapeParser(SpatialContext ctx, SpatialContextFactory factory) {
-    super(ctx,factory);
-  }
+  /**
+   * Write a shape to String
+   */
+  public String toString(Shape shape);
 }
